renamed `dof` directory to `global_assembly`
